refactor(test): Move mocks into their own library
Avoid re-compiling sources for each test that uses them.
