+++ test/mock/include/mock/mockcoresettings.h
