+++ test/model/groupmessagedispatcher_test.cpp
@@ -25,8 +25,8 @@
 #include "src/persistence/igroupsettings.h"
 #include "util/interface.h"
 
-#include "test/mock/mockcoreidhandler.h"
-#include "test/mock/mockgroupquery.h"
+#include "mock/mockcoreidhandler.h"
+#include "mock/mockgroupquery.h"
 
 #include <QObject>
 #include <QtTest/QtTest>
